feat: 차트데이터 firebaseService에 연결, favorite 구조 약간 수정
