+++ app/Public Opinion for Stock/Public Opinion for Stock/Presentation/DetailInfoView/SentimentGraph/SentimentChartView.swift
@@ -5,10 +5,8 @@
 //  Created by 유승재 on 6/2/25.
 //
 
-
 import SwiftUI
 import Charts
-import FirebaseFirestore
 
 extension String {
     func toDate() -> Date? {
