+++ app/Public Opinion for Stock/Public Opinion for Stock/Presentation/DetailInfoView/SentimentGraph/SentimentFrameView.swift
@@ -24,34 +24,7 @@ struct SentimentFrameView: View {
                 Text("오류: \(error)").foregroundColor(.red)
             } else {
                 SentimentChartView(records: viewModel.records)
-                HStack {
-                    HStack(spacing: 16) {
-                        VStack(alignment: .leading) {
-                            Text("긍정적")
-                                .font(.caption)
-                                .foregroundColor(.red)
-                            Text("\(viewModel.records.map { $0.positive }.reduce(0, +))개")
-                                .font(.title3).bold()
-                                .foregroundColor(.red)
-                        }
-                        VStack(alignment: .leading) {
-                            Text("부정적")
-                                .font(.caption)
-                                .foregroundColor(.blue)
-                            Text("\(viewModel.records.map { $0.negative }.reduce(0, +))개")
-                                .font(.title3).bold()
-                                .foregroundColor(.blue)
-                        }
-                        VStack(alignment: .leading) {
-                            Text("중립적")
-                                .font(.caption)
-                                .foregroundColor(.gray)
-                            Text("\(viewModel.records.map { $0.neutral }.reduce(0, +))개")
-                                .font(.title3).bold()
-                                .foregroundColor(.gray)
-                        }
-                    }
-                }
+                SentimentTextView(records: viewModel.records)
             }
             Spacer()
         }
